pyproject.toml
[tool.poetry]
name = "agglomerative_partitioning"
version = "0.1.1"
description = ""
authors = ["Kotaro SONODA <kotaro1976@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
numpy = "^1.22.1"
Pillow = "^9.0.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
ipykernel = "^6.7.0"
matplotlib = "^3.5.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
